// tir-core/src/model/expandableTIR.ts
/**
 * TIRVerifier
 * Specifies the models and endpoints used in Trusted Issuer Registry resolution and issuer verification.
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { Issuer } from './issuer';


/**
 * A potentially incomplete TIR with additional properties for fetching specific issuer\'s data (expanding) in case the TIR uses the \"Tezos\" TIR method, which does not allow fetching all issuers at once.
 */
export interface ExpandableTIR { 
    /**
     * The TIR method identifier, e.g., \"Web\".
     */
    method: any | null;
    /**
     * A constant natural number defining the protocol version of the TIR method. The property enables protocol participants to detect future protocol changes.
     */
    methodProtocol: any | null;
    /**
     * The protocol version of the TIR.
     */
    protocol: any | null;
    /**
     * The DID of the TIR\'s maintainer. This backlink enables serializing the TIR into a Verifiable Credential.
     */
    issuer: any | null;
    /**
     * An ISO 8601 timestamp denoting the last time the TIR was updated. This property allows for smarter caching and thus faster resolution of the TIR. In addition this property becomes the issuanceDate property of the TIR\'s Verifiable Credential.
     */
    lastUpdated: any | null;
    /**
     * Time To Live. The cache expiration in seconds. 
     */
    ttl: any | null;
    /**
     * An object that enables additional metadata-key-value pairs.
     */
    extraMetadata: { [key: string]: any; } | null;
    /**
     * A JSON Map of Issuers
     */
    issuers: { [key: string]: Issuer; } | null;
    /**
     * The DID from which the TIR was resolved. Often, this will match the TIR\'s issuer property.
     */
    did?: string;
    expandedIssuerIds?: Array<string>;
    /**
     * A not further specified object in which a Tezos BigMapAbstraction can be stored. This can be used to fetch individual objects from the Big-map. This is important because the objects of a Tezos Big-map cannot be retrieved all at once, but only individually. To provide the user with this functionality, this property must be passed.
     */
    issuersMichelsonMap?: object;
}