// tir-core/src/model/tIR.ts
/**
 * TIRVerifier
 * Specifies the models and endpoints used in Trusted Issuer Registry resolution and issuer verification.
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { Issuer } from './issuer';


/**
 * A Trusted Issuer Registry containing trusted issuers and metadata describing the TIR.
 */
export interface TIR { 
    /**
     * The TIR method identifier, e.g., \"Web\".
     */
    method: string;
    /**
     * A constant natural number defining the protocol version of the TIR method. The property enables protocol participants to detect future protocol changes.
     */
    methodProtocol: number;
    /**
     * The protocol version of the TIR.
     */
    protocol: number;
    /**
     * The DID of the TIR\'s maintainer. This backlink enables serializing the TIR into a Verifiable Credential.
     */
    issuer: string;
    /**
     * An ISO 8601 timestamp denoting the last time the TIR was updated. This property allows for smarter caching and thus faster resolution of the TIR. In addition this property becomes the issuanceDate property of the TIR\'s Verifiable Credential.
     */
    lastUpdated: string;
    /**
     * Time To Live. The cache expiration in seconds. 
     */
    ttl: number;
    /**
     * An object that enables additional metadata-key-value pairs.
     */
    extraMetadata: { [key: string]: string; } | null;
    /**
     * A JSON Map of Issuers
     */
    issuers: { [key: string]: Issuer; } | null;
}